prayers

+++ utils/handleUploadFile.js
@@ -8,11 +8,20 @@ const removeFile = util.promisify(fs.unlink)
 const handleUploadFile = async (req, res, next) => {
   try {
     const file = req.file;
-    const result = await buckectS3.uploadFile(file);
-    await removeFile(file.path);
-    req.body.file = result.Location;
+    console.log({file});
+    const validType = ["application/pdf", "application/msword", "application/vnd.openxmlformats-officedocument.wordprocessingml.document", "application/vnd.openxmlformats-officedocument.wordprocessingml.template", "application/vnd.ms-powerpoint", "application/vnd.openxmlformats-officedocument.presentationml.presentation", "image/gif", "image/jpeg", "image/png", "application/vnd.ms-excel.sheet.macroEnabled.12"];
 
-    next();
+    if (validType.includes(file.mimetype)) {
+        const result = await buckectS3.uploadFile(file);
+        await removeFile(file.path);
+        req.body.file = result.Location;
+
+        next();
+    } else {
+        const error = new Error();
+        error.message = "invalid file type"
+        throw error;
+    }
   } catch (error) {
     const context = {
       error,
